+++ src/main/java/pl/plajer/villagedefense/arena/initializers/ArenaInitializer1_11_R1.java
@@ -71,7 +71,7 @@ public class ArenaInitializer1_11_R1 extends Arena {
     world.addEntity(fastZombie, CreatureSpawnEvent.SpawnReason.CUSTOM);
     Zombie zombie = (Zombie) fastZombie.getBukkitEntity();
     zombie.setRemoveWhenFarAway(false);
-    CreatureUtils.applyHealthBar(zombie);
+    CreatureUtils.applyHealthAttributes(zombie, this);
     plugin.getHolidayManager().applyHolidayZombieEffects(zombie);
     this.addZombie((Zombie) fastZombie.getBukkitEntity());
 
@@ -88,7 +88,7 @@ public class ArenaInitializer1_11_R1 extends Arena {
     zombie.setRemoveWhenFarAway(false);
     zombie.getEquipment().setBoots(new ItemStack(Material.CHAINMAIL_BOOTS));
     zombie.addPotionEffect(new PotionEffect(PotionEffectType.INVISIBILITY, Integer.MAX_VALUE, 1));
-    CreatureUtils.applyHealthBar(zombie);
+    CreatureUtils.applyHealthAttributes(zombie, this);
     this.addZombie((Zombie) fastZombie.getBukkitEntity());
 
     super.setOptionValue(ArenaOption.ZOMBIES_TO_SPAWN, getOption(ArenaOption.ZOMBIES_TO_SPAWN) - 1);
@@ -108,7 +108,7 @@ public class ArenaInitializer1_11_R1 extends Arena {
     zombie.getEquipment().setChestplate(new ItemStack(Material.DIAMOND_CHESTPLATE));
     zombie.getEquipment().setHelmet(new ItemStack(Material.DIAMOND_HELMET));
     zombie.setRemoveWhenFarAway(false);
-    CreatureUtils.applyHealthBar(zombie);
+    CreatureUtils.applyHealthAttributes(zombie, this);
     this.addZombie((Zombie) fastZombie.getBukkitEntity());
 
     super.setOptionValue(ArenaOption.ZOMBIES_TO_SPAWN, getOption(ArenaOption.ZOMBIES_TO_SPAWN) - 1);
@@ -119,7 +119,7 @@ public class ArenaInitializer1_11_R1 extends Arena {
     BabyZombie fastZombie = new BabyZombie(world);
     fastZombie.setPosition(location.getX(), location.getY(), location.getZ());
     Zombie zombie = (Zombie) fastZombie.getBukkitEntity();
-    CreatureUtils.applyHealthBar(zombie);
+    CreatureUtils.applyHealthAttributes(zombie, this);
     zombie.setRemoveWhenFarAway(false);
     world.addEntity(fastZombie, CreatureSpawnEvent.SpawnReason.CUSTOM);
     plugin.getHolidayManager().applyHolidayZombieEffects(zombie);
@@ -139,7 +139,7 @@ public class ArenaInitializer1_11_R1 extends Arena {
     zombie.getEquipment().setChestplate(new ItemStack(Material.DIAMOND_CHESTPLATE));
     zombie.getEquipment().setHelmet(new ItemStack(Material.DIAMOND_HELMET));
     zombie.setRemoveWhenFarAway(false);
-    CreatureUtils.applyHealthBar(zombie);
+    CreatureUtils.applyHealthAttributes(zombie, this);
     this.addZombie(zombie);
     super.setOptionValue(ArenaOption.ZOMBIES_TO_SPAWN, getOption(ArenaOption.ZOMBIES_TO_SPAWN) - 1);
   }
@@ -156,7 +156,7 @@ public class ArenaInitializer1_11_R1 extends Arena {
     zombie.getEquipment().setChestplate(new ItemStack(Material.IRON_CHESTPLATE));
     zombie.getEquipment().setHelmet(new ItemStack(Material.IRON_HELMET));
     zombie.setRemoveWhenFarAway(false);
-    CreatureUtils.applyHealthBar(zombie);
+    CreatureUtils.applyHealthAttributes(zombie, this);
     this.addZombie(zombie);
     super.setOptionValue(ArenaOption.ZOMBIES_TO_SPAWN, getOption(ArenaOption.ZOMBIES_TO_SPAWN) - 1);
   }
@@ -171,7 +171,7 @@ public class ArenaInitializer1_11_R1 extends Arena {
     zombie.getEquipment().setHelmetDropChance(0.0F);
     zombie.getEquipment().setItemInMainHandDropChance(0F);
     zombie.setRemoveWhenFarAway(false);
-    CreatureUtils.applyHealthBar(zombie);
+    CreatureUtils.applyHealthAttributes(zombie, this);
     this.addZombie(zombie);
 
     super.setOptionValue(ArenaOption.ZOMBIES_TO_SPAWN, getOption(ArenaOption.ZOMBIES_TO_SPAWN) - 1);
@@ -189,7 +189,7 @@ public class ArenaInitializer1_11_R1 extends Arena {
     zombie.getEquipment().setBoots(new ItemStack(Material.GOLD_BOOTS));
     zombie.getEquipment().setLeggings(new ItemStack(Material.GOLD_LEGGINGS));
     zombie.getEquipment().setChestplate(new ItemStack(Material.GOLD_CHESTPLATE));
-    CreatureUtils.applyHealthBar(zombie);
+    CreatureUtils.applyHealthAttributes(zombie, this);
     this.addZombie(zombie);
 
     super.setOptionValue(ArenaOption.ZOMBIES_TO_SPAWN, getOption(ArenaOption.ZOMBIES_TO_SPAWN) - 1);
@@ -207,7 +207,7 @@ public class ArenaInitializer1_11_R1 extends Arena {
     zombie.getEquipment().setLeggings(new ItemStack(Material.CHAINMAIL_LEGGINGS));
     zombie.getEquipment().setChestplate(new ItemStack(Material.CHAINMAIL_CHESTPLATE));
     zombie.getEquipment().setHelmet(new ItemStack(Material.CHAINMAIL_HELMET));
-    CreatureUtils.applyHealthBar(zombie);
+    CreatureUtils.applyHealthAttributes(zombie, this);
     this.addZombie(zombie);
 
     super.setOptionValue(ArenaOption.ZOMBIES_TO_SPAWN, getOption(ArenaOption.ZOMBIES_TO_SPAWN) - 1);